refactor(db): split database.go into logical modules
internal/db/database.go (2,591 lines) split into six focused files
to reduce merge conflicts and improve navigation:

- database.go: core DB connection, Config, sanitizeDBError,
  shared utilities (buildWhereClause, buildUpdateQuery, filterCondition)
- hosts.go: HostRepository, HostFilters, ListHosts, CreateHost,
  GetHost, UpdateHost, DeleteHost, GetHostScans, nullable field helpers
- scans.go: ScanTargetRepository, ScanJobRepository, PortScanRepository,
  Scan/ScanResult/ScanSummary types, ListScans, CreateScan, GetScan,
  UpdateScan, DeleteScan, GetScanResults, GetScanSummary, StartScan,
  StopScan
- networks.go: NetworkSummaryRepository, DiscoveryFilters,
  ListDiscoveryJobs, CreateDiscoveryJob, discovery job stubs
- profiles.go: ProfileFilters, ListProfiles, CreateProfile, GetProfile,
  UpdateProfile, DeleteProfile, parsePostgreSQLArray
- scheduled_jobs.go: Schedule type, ScheduleFilters, ListSchedules,
  CreateSchedule, schedule stubs

No behaviour changes. All existing tests pass, zero linter issues.
