+++ internal/db/networks.go
@@ -1,0 +1,127 @@
+// Package db provides network and discovery-related database operations for scanorama.
+package db
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/google/uuid"
+
+	"github.com/anstrom/scanorama/internal/errors"
+)
+
+// NetworkSummaryRepository handles network summary operations.
+type NetworkSummaryRepository struct {
+	db *DB
+}
+
+// NewNetworkSummaryRepository creates a new network summary repository.
+func NewNetworkSummaryRepository(db *DB) *NetworkSummaryRepository {
+	return &NetworkSummaryRepository{db: db}
+}
+
+// GetAll retrieves all network summaries.
+func (r *NetworkSummaryRepository) GetAll(ctx context.Context) ([]*NetworkSummary, error) {
+	var summaries []*NetworkSummary
+	query := `
+		SELECT *
+		FROM network_summary
+		ORDER BY target_name`
+
+	if err := r.db.SelectContext(ctx, &summaries, query); err != nil {
+		return nil, sanitizeDBError("get network summaries", err)
+	}
+
+	return summaries, nil
+}
+
+// DiscoveryFilters represents filters for listing discovery jobs.
+type DiscoveryFilters struct {
+	Status string
+	Method string
+}
+
+// ListDiscoveryJobs retrieves discovery jobs with filtering and pagination.
+func (db *DB) ListDiscoveryJobs(
+	ctx context.Context,
+	filters DiscoveryFilters,
+	offset, limit int,
+) ([]*DiscoveryJob, int64, error) {
+	jobs := []*DiscoveryJob{}
+	total := int64(0)
+	return jobs, total, nil
+}
+
+// CreateDiscoveryJob creates a new discovery job.
+func (db *DB) CreateDiscoveryJob(ctx context.Context, jobData interface{}) (*DiscoveryJob, error) {
+	data, ok := jobData.(map[string]interface{})
+	if !ok {
+		return nil, fmt.Errorf("invalid discovery job data format")
+	}
+
+	// Extract data from request.
+	networks, ok := data["networks"].([]string)
+	if !ok || len(networks) == 0 {
+		return nil, fmt.Errorf("networks are required and must be a string array")
+	}
+
+	method := data["method"].(string)
+	if method == "" {
+		method = DiscoveryMethodTCP
+	}
+
+	// For simplicity, create one discovery job for the first network.
+	// In a production system, you might create multiple jobs or handle multiple networks differently.
+	network := networks[0]
+
+	jobID := uuid.New()
+	now := time.Now().UTC()
+
+	query := `
+		INSERT INTO discovery_jobs (id, network, method, status, created_at, hosts_discovered, hosts_responsive)
+		VALUES ($1, $2, $3, 'pending', $4, 0, 0)
+	`
+
+	_, err := db.ExecContext(ctx, query, jobID, network, method, now)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create discovery job: %w", err)
+	}
+
+	job := &DiscoveryJob{
+		ID:              jobID,
+		Network:         NetworkAddr{}, // Would need proper parsing.
+		Method:          method,
+		Status:          "pending",
+		CreatedAt:       now,
+		HostsDiscovered: 0,
+		HostsResponsive: 0,
+	}
+
+	return job, nil
+}
+
+// GetDiscoveryJob retrieves a discovery job by ID.
+func (db *DB) GetDiscoveryJob(ctx context.Context, id uuid.UUID) (*DiscoveryJob, error) {
+	return nil, errors.ErrNotFoundWithID("discovery job", id.String())
+}
+
+// UpdateDiscoveryJob updates an existing discovery job.
+func (db *DB) UpdateDiscoveryJob(ctx context.Context, id uuid.UUID, jobData interface{}) (*DiscoveryJob, error) {
+	return nil, errors.ErrNotFoundWithID("discovery job", id.String())
+}
+
+// DeleteDiscoveryJob deletes a discovery job by ID.
+func (db *DB) DeleteDiscoveryJob(ctx context.Context, id uuid.UUID) error {
+	return errors.ErrNotFoundWithID("discovery job", id.String())
+}
+
+// StartDiscoveryJob starts discovery job execution.
+func (db *DB) StartDiscoveryJob(ctx context.Context, id uuid.UUID) error {
+	return errors.ErrNotFoundWithID("discovery job", id.String())
+}
+
+// StopDiscoveryJob stops discovery job execution.
+func (db *DB) StopDiscoveryJob(ctx context.Context, id uuid.UUID) error {
+	return errors.ErrNotFoundWithID("discovery job", id.String())
+}
