+++ internal/db/scheduled_jobs.go
@@ -1,0 +1,86 @@
+// Package db provides scheduled job database operations for scanorama.
+package db
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/google/uuid"
+
+	"github.com/anstrom/scanorama/internal/errors"
+)
+
+// Schedule represents a scheduled task.
+type Schedule struct {
+	ID             uuid.UUID              `json:"id" db:"id"`
+	Name           string                 `json:"name" db:"name"`
+	Description    string                 `json:"description" db:"description"`
+	CronExpression string                 `json:"cron_expression" db:"cron_expression"`
+	JobType        string                 `json:"job_type" db:"job_type"`
+	JobConfig      map[string]interface{} `json:"job_config" db:"job_config"`
+	Enabled        bool                   `json:"enabled" db:"enabled"`
+	CreatedAt      time.Time              `json:"created_at" db:"created_at"`
+	UpdatedAt      time.Time              `json:"updated_at" db:"updated_at"`
+	LastRun        *time.Time             `json:"last_run" db:"last_run"`
+	NextRun        *time.Time             `json:"next_run" db:"next_run"`
+}
+
+// ScheduleFilters represents filters for listing schedules.
+type ScheduleFilters struct {
+	Enabled bool
+	JobType string
+}
+
+// ListSchedules retrieves schedules with filtering and pagination.
+func (db *DB) ListSchedules(
+	ctx context.Context,
+	filters ScheduleFilters,
+	offset, limit int,
+) ([]*Schedule, int64, error) {
+	schedules := []*Schedule{}
+	total := int64(0)
+	return schedules, total, nil
+}
+
+// CreateSchedule creates a new schedule.
+func (db *DB) CreateSchedule(ctx context.Context, scheduleData interface{}) (*Schedule, error) {
+	_, ok := scheduleData.(map[string]interface{})
+	if !ok {
+		return nil, fmt.Errorf("invalid schedule data format")
+	}
+
+	schedule := &Schedule{
+		ID:        uuid.New(),
+		Enabled:   true,
+		CreatedAt: time.Now().UTC(),
+		UpdatedAt: time.Now().UTC(),
+	}
+
+	return schedule, nil
+}
+
+// GetSchedule retrieves a schedule by ID.
+func (db *DB) GetSchedule(ctx context.Context, id uuid.UUID) (*Schedule, error) {
+	return nil, errors.ErrNotFoundWithID("schedule", id.String())
+}
+
+// UpdateSchedule updates an existing schedule.
+func (db *DB) UpdateSchedule(ctx context.Context, id uuid.UUID, scheduleData interface{}) (*Schedule, error) {
+	return nil, errors.ErrNotFoundWithID("schedule", id.String())
+}
+
+// DeleteSchedule deletes a schedule by ID.
+func (db *DB) DeleteSchedule(ctx context.Context, id uuid.UUID) error {
+	return errors.ErrNotFoundWithID("schedule", id.String())
+}
+
+// EnableSchedule enables a schedule.
+func (db *DB) EnableSchedule(ctx context.Context, id uuid.UUID) error {
+	return errors.ErrNotFoundWithID("schedule", id.String())
+}
+
+// DisableSchedule disables a schedule.
+func (db *DB) DisableSchedule(ctx context.Context, id uuid.UUID) error {
+	return errors.ErrNotFoundWithID("schedule", id.String())
+}
